refactor: split settings into src/settings/
Break getters, validation, NVS, load, save, and reset into separate files so the old 555-line storage module is easier to follow.
