+++ src/sleep.cpp
@@ -9,7 +9,7 @@
 #include "network/wifi_connect.h"
 #include "pins.h"
 #include "servos.h"
-#include "settings.h"
+#include "settings/settings.h"
 #include "sleep.h"
 
 namespace {
